requirements.txt
# Основные зависимости
aiohttp==3.9.1
requests==2.31.0
asyncio==3.4.3
websockets==12.0
sqlite3

# Конфигурация и переменные окружения
python-dotenv==1.0.0
PyYAML==6.0.2

# Кэширование и производительность
cachetools==5.3.2
redis==5.0.1

# Веб-интерфейс
Flask==3.0.0
Flask-CORS==4.0.0
gunicorn==21.2.0

# Аналитика и визуализация
pandas==2.1.4
matplotlib==3.8.2
numpy==1.24.3
plotly==5.17.0

# Машинное обучение
scikit-learn==1.3.2
scipy==1.11.4

# Социальные сети
telethon==1.32.1
discord.py-self==1.9.2
tweepy==4.14.0
snscrape==0.7.0.20230622

# AI и анализ текста
openai==1.3.7
transformers==4.36.2
torch==2.1.2
nltk==3.8.1
textblob==0.17.1

# Парсинг и обработка данных
beautifulsoup4==4.12.2
lxml==4.9.3
feedparser==6.0.10
selenium==4.15.2
playwright==1.40.0

# Безопасность и шифрование
cryptography==41.0.8
bcrypt==4.1.2

# Логирование и мониторинг
structlog==23.2.0
prometheus-client==0.19.0

# Тестирование
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Утилиты
click==8.1.7
rich==13.7.0
tqdm==4.66.1
colorama==0.4.6

# Дополнительные API клиенты
ccxt==4.1.77
defillama==0.0.1a0
dexscreener==0.1.0

# Системные зависимости
psutil==5.9.6
py-cpuinfo==9.0.0

# Документация
sphinx==7.2.6
sphinx-rtd-theme==1.3.0 